requirements.txt
langgraph
langchain
langchain-groq
langchain-community
python-dotenv
docker
pydantic
termcolor
tavily-python